Day-4(Leetcode-Opti)/requirements.txt
streamlit>=1.28.0
pyautogen>=0.2.0
groq>=0.4.0
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.31.0
